mlir/utils/migraphxRunner/requirements.txt
setuptools
PyYAML==5.4
